fix: move the getProductKey to his own file
